abstracting camera hardware
