+++ src/hardware/camera/rpi_camera.py
@@ -1,0 +1,21 @@
+import numpy as np
+from picamera2 import Picamera2
+from src.utils.logging import logging_default
+
+class RPiCamera():
+    def __init__(self):
+        self.picam2 = Picamera2()
+        self.picam2.start()
+
+    def get_capture(self) -> np.ndarray:
+        try:
+            frame = self.picam2.capture_array()
+            if frame is not None:
+                return True, frame
+            logging_default.warning("PiCamera2 returned None")
+        except Exception as e:
+            logging_default.error(f"Error capturing frame image from Pi Camera 2: {e}")
+        return False, None
+
+    def release(self):
+        pass
